Move lots of AST constructors to headers
---
