+++ compiler/next/lib/uast/Variable.cpp
@@ -25,27 +25,6 @@ namespace chpl {
 namespace uast {
 
 
-Variable::Variable(ASTList children,
-                   UniqueString name, Sym::Visibility vis,
-                   Variable::Tag tag,
-                   int8_t typeExpressionChildNum,
-                   int8_t initExpressionChildNum)
-  : Sym(asttags::Variable, std::move(children), name, vis),
-    tag_(tag),
-    typeExpressionChildNum(typeExpressionChildNum),
-    initExpressionChildNum(initExpressionChildNum) {
-
-  assert(-1 <= typeExpressionChildNum && typeExpressionChildNum <= 1);
-  assert(-1 <= initExpressionChildNum && initExpressionChildNum <= 1);
-  assert(numChildren() <= 2);
-#ifndef NDEBUG
-  // check that all children are exprs (and not, say, Syms)
-  for (const ASTNode* child : this->children()) {
-    assert(child->isExpression());
-  }
-#endif
-}
-
 bool Variable::contentsMatchInner(const ASTNode* other) const {
   const Variable* lhs = this;
   const Variable* rhs = (const Variable*) other;
